Alternative implementation to combine the API endpoints
This is an alternative implementation of how to handle the custom-images
and custom-images/imports endpoints. It only overwrites the _get, _post
and present methods and has less code duplication from
AnsibleCloudscaleBase.

The _get_url method can be removed once the API bug is fixed.

It adds an additional parameter "force_retry" to retry an image import
even if a failed import using the same name and URL already exists. This
is necessary for cases where the failure is not because the wrong
URL was used but because something else failed (eg. URL temporarily not
working).