Fix failing load balancer tests
The load balancer tests verifying that a task referencing an inexistent
load balancer fails got broken because of a change in the API response.

This fixes the tests to match part the new API response body. It does
not match the whole body to reduce the chance of the test breaking again
if the API response slightly changes.