Validate API token before passing it on to Ansible
When the token is set to a value that is not a valid HTTP header,
Ansible's http client  rejects it with an error that reveals the token.

This may cause it to be accidentally logged.

We now validate the token, to ensure it is a strict subset of valid
HTTP header values. It includes a bit more characters than we currently
use for our tokens, just in case we change our format.